# src/build/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "REST API proyectoUTM",
        "description": "<html><body>Esta es la documentación de la API para el sistema de Investigación Docencia.<p>Para poder usar esta API es necesario solicitar el token y con esto poder utilizar los servicios proporcionados.</p><br><p><b>Autenticación y token de API</b></p><p>La Universidad Tecnológica de la Mixteca (UTM) proporciona el api-token. Una vez que haya creado y confirmado su cuenta.</p></body></html>",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:3000",
    "basePath": "",
...
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/api/institutos/eliminar/{idInstituto}": {
            "delete": {
                "summary": "Eliminar un instituto",
                "operationId": "eliminar",
                "parameters": [
                    {
                        "name": "idInstituto",
                        "in": "path",
                        "description": "id del Instituto a eliminar",
                        "required": true,
                        "default": 14
                    }
                ],
                "tags": [
                    "Institutos"
                ],
                "responses": {
                    "201": {
                        "description": "respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
        "/api/profesores/eliminar/{idProfesor}": {
            "delete": {
                "summary": "Eliminar un registro de un profesor",
                "operationId": "eliminar",
                "parameters": [
                    {
                        "name": "idProfesor",
                        "in": "path",
                        "description": "identificador del profesor a eliminar",
                        "required": true,
                        "default": 14
                    }
                ],
                "tags": [
                    "Profesores"
                ],
                "responses": {
                    "201": {
                        "description": "respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
        "/api/carreras/eliminar/{idCarrera}": {
            "delete": {
                "summary": "Eliminar una carrera",
                "operationId": "eliminar",
                "parameters": [
                    {
                        "name": "idCarrera",
                        "in": "path",
                        "description": "Identificador de la carrera a eliminar",
                        "required": true,
                        "default": 14
                    }
                ],
                "tags": [
                    "Carreras"
                ],
                "responses": {
                    "201": {
                        "description": "respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
        "/api/tipoProfesor/eliminar/{idTipoProfesor}": {
            "delete": {
                "summary": "Eliminar un tipo de proesor",
                "operationId": "eliminar",
                "parameters": [
                    {
                        "name": "idTipoProfesor",
                        "in": "path",
                        "description": "Identificador del tipo profesor a eliminar",
                        "required": true,
                        "default": 14
                    }
                ],
                "tags": [
                    "tipos de Profesores"
                ],
                "responses": {
                    "201": {
                        "description": "respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
...
        "/api/institutos/actualizar/{idInstituto}": {
            "put": {
                "summary": "Modificando un instituto",
                "operationId": "actualizar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "idInstituto",
                        "in": "path",
                        "description": "id del Instituto a eliminar",
                        "required": true,
                        "default": 18
                    },
                    {
                        "name": "Instituto",
                        "in": "body",
                        "description": "Instituto a modificar",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ViejoInstituto"
                        }
                    }
                ],
                "tags": [
                    "Institutos"
                ],
                "responses": {
                    "201": {
                        "description": "respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
        "/api/profesores/actualizar/{idProfesor}": {
            "put": {
                "summary": "Modificando un registro profesor",
                "operationId": "actualizar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "idProfesor",
                        "in": "path",
                        "description": "identificador del profesor a eliminar",
                        "required": true,
                        "default": 18
                    },
                    {
                        "name": "Profesor",
                        "in": "body",
                        "description": "Profesor a modificar",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ViejoProfesor"
                        }
                    }
                ],
                "tags": [
                    "Profesores"
                ],
                "responses": {
                    "201": {
                        "description": "respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
        "/api/carreras/actualizar/{idCarrera}": {
            "put": {
                "summary": "Modificando una carrera",
                "operationId": "actualizar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "idCarrera",
                        "in": "path",
                        "description": "Identificador de la carrera a actualizar",
                        "required": true,
                        "default": 8
                    },
                    {
                        "name": "Carreras",
                        "in": "body",
                        "description": "Carrera a modificar",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ViejoCarrera"
                        }
                    }
                ],
                "tags": [
                    "Carreras"
                ],
                "responses": {
                    "201": {
                        "description": "respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
        "/api/tipoProfesor/actualizar/{idTipoProfesor}": {
            "put": {
                "summary": "Modificando un tipo profesor",
                "operationId": "actualizar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "idTipoProfesor",
                        "in": "path",
                        "description": "identificador del tipo profesor a eliminar",
                        "required": true,
                        "default": 18
                    },
                    {
                        "name": "TipoProfesor",
                        "in": "body",
                        "description": "tipo profesor a modificar",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ViejoTipoProfesor"
                        }
                    }
                ],
                "tags": [
                    "tipos de Profesores"
                ],
                "responses": {
                    "201": {
                        "description": "respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "ViejoInstituto": {
            "type": "object",
            "required": [
                "codigo",
                "nombre"
            ],
            "properties": {
                "codigo": {
                    "type": "string",
                    "default": "1"
                },
                "nombre": {
                    "type": "string",
                    "default": "Instituto prueba"
                }
            }
        },
        "ViejoTipoProfesor": {
            "type": "object",
            "required": [
                "idTipoProfesor",
                "nombreTipo"
            ],
            "properties": {
                "codigo": {
                    "type": "number",
                    "default": "1"
                },
                "nombre": {
                    "type": "string",
                    "default": "prueba"
                }
            }
        },
        "ViejoCarrera": {
            "type": "object",
            "required": [
                "codigoCarrera",
                "nombre",
                "idInstituto"
            ],
            "properties": {
                "codigoCarrera": {
                    "type": "string",
                    "default": "15"
                },
                "nombre": {
                    "type": "string",
                    "default": "Instituto prueba"
                },
                "idInstituto": {
                    "type": "number",
                    "default": "13"
                }
            }
        },
        "ViejoProfesor": {
            "type": "object",
            "required": [
                "nombres",
                "apellidoPaterno",
                "apellidoMaterno",
                "correoProfesor",
                "password",
                "nivel",
                "idCarrera",
                "grado",
                "idTipoProfesor"
            ],
            "properties": {
                "nombres": {
                    "type": "string",
                    "default": "Prueba"
                },
                "apellidoPaterno": {
                    "type": "string",
                    "default": "prueba2"
                },
                "apellidoMaterno": {
                    "type": "string",
                    "default": "prueba3"
                },
                "correoProfesor": {
                    "type": "string",
                    "default": "prueba4"
                },
                "password": {
                    "type": "string",
                    "default": "adsdas"
                },
                "nivel": {
                    "type": "number",
                    "default": "2"
                },
                "idCarrera": {
                    "type": "number",
                    "default": "1"
                },
                "grado": {
                    "type": "string",
                    "default": "2"
                },
                "idTipoProfesor": {
                    "type": "number",
                    "default": "1"
                }
            }
        },
...
    }
}